Create the tabsHelpers module
